extract_and_load/requirements.txt
appdirs==1.4.4
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
boto3==1.28.2
botocore==1.31.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cramjam==2.6.2
exceptiongroup==1.1.2
fastparquet==2023.7.0
Flask==2.3.2
fsspec==2023.6.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
MarkupSafe==2.1.3
nfl-data-py==0.3.0
numpy==1.24.4
outcome==1.2.0
packaging==23.1
pandas==2.0.3
pyarrow==12.0.1
PySocks==1.7.1
python-dateutil==2.8.2
python-snappy==0.6.1
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.6.1
scikit-learn==1.3.0
scipy==1.10.1
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
threadpoolctl==3.2.0
trio==0.22.2
trio-websocket==0.10.3
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6
wsproto==1.2.0
xgboost==1.7.6
zipp==3.16.2
